+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixes
 
 - Fix processor architecture for Linux ([#355](https://github.com/getsentry/sentry-dart-plugin/pull/355))
+- Add custom `symbols_path` in iOS symbol map discovery ([#394](https://github.com/getsentry/sentry-dart-plugin/pull/394))
 
 ## 3.2.1